renamed common folder to corelib
