+++ examples/lib-polling/lib-polling.ino
@@ -4,17 +4,10 @@
 
 
 #ifdef XMC1100_XMC2GO
+#define Q_OUTPUT_PIN    9   /**< Hall Switch Output Pin  */
 #define LED1 14             /**< LED1 Pin Allocation for XMC2Go  */
 #else
 #define LED1 LED_BUILTIN
-#endif
-
-#ifdef XMC1100_XMC2GO 
-#define Q_OUTPUT_PIN    9   /**< Hall Switch Output Pin  */
-#elif ((XMC1100_Boot_Kit) ||  \
-      (XMC4700_Relax_Kit) ||  \
-      (ARDUINO_AVR_UNO))
-#define Q_OUTPUT_PIN    3   /**< Hall Switch Output Pin  */
 #endif
 
 HallSwitchIno        hs(Q_OUTPUT_PIN);  /**< Hall Switch object */
