+++ src/corelib/hall-conf-dfl.h
@@ -27,6 +27,4 @@
 
 #define HALL_SWITCH_FRAMEWORK   HALL_SWITCH_FRAMEWORK_ARDUINO
 
-
-
 #endif /** HALL_CONF_DFL_H_ **/
