+++ src/corelib/hall-speed.cpp
