+++ src/corelib/hall-speed.h
