+++ src/corelib/hall-switch-int.cpp
