+++ src/corelib/hall-switch-int.h
