+++ src/corelib/hall-switch.cpp
