+++ src/corelib/hall-switch.h
